+++ src/index.ts
@@ -21,4 +21,4 @@ export type {
 } from "./data/types";
 export type { Resolution } from "./data";
 
-export const VERSION = "0.16.0";
+export const VERSION = "0.17.0";